requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
motor
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
requests
boto3
pytest
httpx
pytest-asyncio
pymongo
redis